Change isActive to isEditing and move from ReactDataGrid state to RxJsState
