+++ packages/react-data-grid/src/RowComparer.js
@@ -1,13 +1,5 @@
 import ColumnMetrics from './ColumnMetrics';
 
-function doesRowContainSelectedCell(props) {
-  let selected = props.cellMetaData.selected;
-  if (selected && selected.rowIdx === props.idx) {
-    return true;
-  }
-  return false;
-}
-
 function willRowBeDraggedOver(props) {
   let dragged = props.cellMetaData.dragged;
   return dragged != null && (dragged.rowIdx >= 0 || dragged.complete === true);
@@ -20,8 +12,6 @@ function hasRowBeenCopied(props) {
 
 export const shouldRowUpdate = (nextProps, currentProps) => {
   return !(ColumnMetrics.sameColumns(currentProps.columns, nextProps.columns, ColumnMetrics.sameColumn)) ||
-    doesRowContainSelectedCell(currentProps) ||
-    doesRowContainSelectedCell(nextProps) ||
     willRowBeDraggedOver(nextProps) ||
     nextProps.row !== currentProps.row ||
     currentProps.colDisplayStart !== nextProps.colDisplayStart ||
