+++ packages/react-data-grid/src/stateManagement/state/CellState.js
@@ -1,0 +1,11 @@
+export default {
+  selectedPosition: {
+    idx: 0,
+    rowIdx: 0
+  },
+  lastSelectedPosition: {
+    idx: 0,
+    rowIdx: 0
+  },
+  isEditing: false
+};
