+++ packages/react-data-grid/src/stateManagement/state/InitialState.js
@@ -1,0 +1,5 @@
+import cell from './CellState';
+
+export default {
+  ...{cell}
+};
